read rfid data

+++ android/app/src/main/kotlin/com/luojie/erapp/inventory_app/data/Rfid.kt
@@ -11,4 +11,4 @@ import android.bluetooth.BluetoothDevice
 
 
 
-data class BtDeviceInfo(val macAddress : BluetoothDevice, val name: String, val rssi: Int,var  isSelected : Boolean )
+data class BtDeviceInfo(val macAddress : BluetoothDevice, val name: String, val rssi: Short,var  isSelected : Boolean )
